+++ template/src/modules/main.js
@@ -14,6 +14,7 @@ Object.keys(filters).forEach((key) => {
   Vue.filter(key, filters[key]);
 });
 
+/* eslint-disable no-new */
 new Vue({
   el: '#app',
   router,