+++ template/src/modules/services/index.js
@@ -9,5 +9,5 @@
 import _Axios from '../utils/_axios';
 
 export function login(data) {
-    return _Axios('/usercenter/user/mobile_login', { data });
+  return _Axios('/usercenter/user/mobile_login', {data});
 }